[CIR][Dialect] Add `convergent` attribute to functions for SIMT languages (#840)
Fix #805. This PR includes end-to-end implementation.

The `convergent` attribute is set depending on languages, which is
wrapped as `langOpts.assumeFunctionsAreConvergent()`.

Therefore, in ClangIR, every `cir.func` under `#cir.lang<opencl_c>` is
set to be convergent.

After lowering to LLVM IR, `PostOrderFunctionAttrs` pass will remove
unnecessary `convergent` then. In other words, we will still see
`convergent` on every function with `-O0`, but not with default
optimization level.

The test taken from `CodeGenOpenCL/convergent.cl` is a bit complicated.
However, the core of it is that `convergent` is set properly for
`convfun()` `non_convfun()` `f()` and `g()`. Merge of two `if` is more
or less a result of generating the same LLVM IR as OG.