[CIR] Split cir-simplify into two passes
This patch splits the old cir-simplify pass into two new passes, namely
cir-canonicalize and cir-simplify (the new cir-simplify). The cir-canonicalize
pass runs transformations that do not affect CIR-to-source fidelity much, such
as operation folding and redundant operation elimination. On the other hand, the
new cir-simplify pass runs transformations that may significantly change the
code and break high-level code analysis passes, such as more aggresive code
optimizations.

This patch also updates the CIR-to-CIR pipeline to fit these two new passes. The
cir-canonicalize pass is moved to the very front of the pipeline, while the new
cir-simplify pass is moved to the back of the pipeline (but still before
lowering prepare of course). Additionally, the new cir-simplify now only runs
when the user specifies a non-zero optimization level on the frontend.