[CIR][CodeGen][NFC] Rename the confusing `buildGlobal` overload (#897)
`CIRGenModule::buildGlobal` --[rename]-->
`CIRGenModule::getOrCreateCIRGlobal`

We already have `CIRGenModule::buildGlobal` that corresponds to
`CodeGenModule::EmitGlobal`. But there is an overload of `buildGlobal`
used by `getAddrOfGlobalVar`. Since this name is confusing, this PR
rename it to `getOrCreateCIRGlobal`.

Note that `getOrCreateCIRGlobal` already exists. It is intentional to
make the renamed function an overload to it. The reason here is that the
renamed function is basically a wrapper of the original
`getOrCreateCIRGlobal` with more specific parameters:

`getOrCreateCIRGlobal(decl, type, isDef)` --[call]-->
`getOrCreateCIRGlobal(getMangledName(decl), type,
decl->getType()->getAS(), decl, isDef)`