[CIR][CodeGen] kr-style for function arguments (#938)
I tried to run llvm-test-suite and turned out that there are many tests
fail with segfault due to old C style (let's remember Kernighan and
Ritchie) . This PR fix it by the usual copy-pasta from the original
codegen :)

So let's take a look at the code:
```
void foo(x) short x; {}
int main() {
  foo(4); 
  return 0;
}
```
and CIR for `foo` function is:
```
cir.func  @foo(%arg0: !s32i) {
    %0 = cir.alloca !s16i, !cir.ptr<!s16i>, ["x", init] 
    %1 = cir.cast(bitcast, %0 : !cir.ptr<!s16i>), !cir.ptr<!s32i>
    cir.store %arg0, %1 : !s32i, !cir.ptr<!s32i>
    cir.return
}
```
We bitcast the **address** (!!!) and store a value of a bigger size
there.

And now everything looks fine:
```
cir.func no_proto  @foo(%arg0: !s32i) {
    %0 = cir.alloca !s16i, !cir.ptr<!s16i>, ["x", init] 
    %1 = cir.cast(integral, %arg0 : !s32i), !s16i
    cir.store %1, %0 : !s16i, !cir.ptr<!s16i> 
    cir.return 
} 
```
We truncate an argument and store it. 

P.S.
The `bitcast` that was there before looks a little bit suspicious and
dangerous. Are we sure we can do this unconditional cast while we create
`StoreOp` ?