[CIR][CIRGen] Add const attribute to alloca operations (#892)
This PR tries to give a simple initial implementation for eliminating
redundant loads of constant objects, an idea originally posted by
OfekShilon.

Specifically, this PR adds a new unit attribute `const` to the
`cir.alloca` operation. Presence of this attribute indicates that the
alloca-ed object is declared `const` in the input source program. CIRGen
is updated accordingly to start emitting this new attribute.