[CIR][Lowering] Transform cir.store of const arrays into cir.copy
Add lowering prepare logic to lower stores to cir.copy.

This bring LLVM lowering closer to OG and turns out the rest of the compiler
understands memcpys better and generate better assembly code for at least arm64
and x86_64.

Note that current lowering to memcpy is only using i32 intrinsic version,
this PR does not touch that code and that will be addressed in another PR.