[CIR] Add a cir.undef attr
CodeGen sometimes emits undef constants directly, e.g. when initializing
an empty struct (https://godbolt.org/z/68od33aa8). We want to match this
behavior, so we need a cir.undef attr to represent the constant. This
change implements the lowering for the new op, which matches how
cir.zero is lowered. A follow-up will change CIRGen to use it. It also
replaces UndefOf with a ConstantOp of UndefAttr to avoid redundancy.

Pull Request resolved: https://github.com/llvm/clangir/pull/993