[ClangIR][CIRGen] Handle nested union in arrays of struct (#1007)
Reproducer:

```
struct nested
{
  union {
    const char *single;
    const char *const *multi;
  } output;
};
static const char * const test[] = {
  "test",
};
const struct nested data[] = 
{
    {
        {
            .multi = test,
        },
    },
    {
        {
            .single = "hello",
        },
    },
};
```

ClangIR now failed to recognize `data` as an array since it failed to
recognize the initializer for union. This comes from a fundamental
difference between CIR and LLVM IR. In LLVM IR, the union is simply a
struct with the largest member. So it is fine to have only one init
element. But in CIR, the union has the information for all members. So
if we only pass a single init element, we may be in trouble. We solve
the problem by appending placeholder attribute for the uninitialized
fields.