[CIR][CIRGen] Add aliases for virtual dtors calls
Note that we lack two pieces of support for aliases in LLVM IR dialect
globals: the `alias` keyword and function types `void (ptr)`, this needs
to be done before we can nail this for good, but it's outside the scope
of this commit.

The behavior is slightly different under -O1, which will be addressed next.