[CIR] Extend support for floating point attributes (#572)
This commit extends the support for floating point attributes parsing by
using
the new `AsmParser::parseFloat(fltSemnatics, APFloat&)` interface.
As a drive-by, this commit also harmonizes the cir.fp print/parse
namespace
usage, and adds the constraint of supporting only "CIRFPType"s for
cir.fp in
tablegen instead of verifying it manually in the parsing logic.

---

This commit is based on top of a to-be-upstreamed commit which extends
the upstream MLIR float type parsing. Upstream parsing of float type has
full capability only through parsing the Builtin Dialect's `FloatAttr`.
Thos commit exposes the same capabilities to downstream users.

---

This PR should resolve (at least) `GCC-C-execute-ieee-fp-cmp-2` and
`GCC-C-execute-ieee-fp-cmp-4`, paving the way to other
`GCC-C-execute-ieee-*` tests passing from the SingleSource suite. It
resolves #559 .