[CIR][CodeGen][NFC] Simplify replacing initializer of static decls (#1095)
When emitting initializers for static declarations, it's essential to
ensure that the `cir.global` operation aligns its type with that of the
initializer.

The original approach creates a new global op and copies every attribute
from the old one. But just `setSymType` should work well.

This also removes missing feature flags there.