[CIR] Properly ensure terminating IfOp and ScopeOp regions (#1097)
The code changes modify the `cir.if` and `cir.scope` operations to
ensure that their code regions are properly terminated. Previously, the
if/else and scope regions could be left completely empty which is
non-trivially expected in most code inspecting these ops. This led, for
example, to a crash when and if clause was left empty in the source
code.

Now, the child regions must be terminated, either explicitly or
implicitly by the default builder and assembly parser. This change
improves the clarity and correctness of the code.