[CIR] [Lowering] [X86_64] Support VAArg for LongDouble (#1101)
This is the following of https://github.com/llvm/clangir/pull/1100.

After https://github.com/llvm/clangir/pull/1100, when we want to use
LongDouble for VAArg, we will be in trouble due to details in X86_64's
ABI and this patch tries to address this.

The practical impact the patch is, after this patch, with
https://github.com/llvm/clangir/pull/1088 and a small following up fix,
we can build and run all C's benchmark in SpecCPU 2017. I think it is a
milestone.