[CIR][ABI][AArch64] Support struct passing with coercion through memory  (#1111)
This PR adds a support for one more case of passing structs by value,
with `memcpy` emitted.

First of all, don't worry - despite the PR seems big, it's basically
consist of helpers + refactoring. Also, there is a minor change in the
`CIRBaseBuilder` - I made static the `getBestAllocaInsertPoint` method
in order to call it from lowering - we discussed once - and I here we
just need it (or copy-paste the code, which doesn't seem good).

I will add several comments in order to simplify review.