[CIR] Make use of !invariant.group metadata for const allocas
This patch updates the LLVM lowering part of load and stores to const allocas
and makes use of the !invariant.group metadata in the result LLVM IR.

The HoistAlloca pass is also updated. The const flag on a hoisted alloca is
removed for now since their uses are not always invariants. Will update in later
patches to teach their invariants.