[CIR] [Lowering] [X86_64] Support VAArg for LongDouble (#1150)
Recommit https://github.com/llvm/clangir/pull/1101

I am not sure what happened. But that merged PR doesn't show in the git
log. Maybe the stacked PR may not get successed? But after all, we need
to land it again.

Following off are original commit messages:

---

This is the following of https://github.com/llvm/clangir/pull/1100.

After https://github.com/llvm/clangir/pull/1100, when we want to use
LongDouble for VAArg, we will be in trouble due to details in X86_64's
ABI and this patch tries to address this.

The practical impact the patch is, after this patch, with
https://github.com/llvm/clangir/pull/1088 and a small following up fix,
we can build and run all C's benchmark in SpecCPU 2017. I think it is a
milestone.