[CIR][Dialect][NFC] Remove redundant module attribute `llvm.data_layout` (#1156)
Since LLVM specific data layout string is not proper in ClangIR, this PR
replaces it with existing MLIR DLTI equivalent and eliminate the
redundancy.

Although the constructor of `LowerModule` of TargetLowering library
requires a llvm data layout string, it is not used currently. (I believe
it would also not matter in the future.) Therefore, this PR has no
functional change.