[CIR][ABI][AArch64][Lowering] Fix the callsite for nested unions (#1169)
For example, the following reaches
["NYI"](https://github.com/llvm/clangir/blob/c8b626d49e7f306052b2e6d3ce60b1f689d37cb5/clang/lib/CIR/Dialect/Transforms/TargetLowering/LowerFunction.cpp#L348)
when lowering to AArch64:
```
typedef struct {
  union {
    struct {
      char a, b;
    };
    char c;
  };
} A;

void foo(A a) {}

void bar() {
  A a;
  foo(a);
}
```
Currently, the value of the struct becomes a bitcast operation, so we
can simply extend `findAlloca` to be able to trace the source alloca
properly, then use that for the
[coercion](https://github.com/llvm/clangir/blob/c8b626d49e7f306052b2e6d3ce60b1f689d37cb5/clang/lib/CIR/Dialect/Transforms/TargetLowering/LowerFunction.cpp#L341)
through memory. I have also added a test for this case.