[CIR] Remove InavalidUnionFieldAttr and remove a check in array attr (#1166)
Close https://github.com/llvm/clangir/issues/1131

This is another solution to https://github.com/llvm/clangir/pull/1160

This patch revert https://github.com/llvm/clangir/pull/1007 and remain
its test. The problem described in
https://github.com/llvm/clangir/pull/1007 is workaround by skipping the
check of equivalent of element types in arrays.

We can't mock such checks simply by adding another attribute to
`ConstStructAttr` since the types are aggregated. e.g., we have to
handle the cases like `struct { union { ... } }` and `struct { struct {
union { ... } } }` and so on. To make it, we have to introduce what I
called "two type systems" in https://github.com/llvm/clangir/pull/1160.

This is not very good giving it removes a reasonable check. But it might
not be so problematic since the Sema part has already checked it. (Of
course, we still need face the risks to introduce new bugs any way)