[CIR] Add code generation options to lowering context
This PR adds `clang::CodeGenOptions` to the lowering context. Similar to
`clang::LangOptions`, the code generation options are currently set to the
default values when initializing the lowering context.

Besides, this PR also adds a new attribute `#cir.opt_info`. The attribute is a
module-level attribute and it holds optimization information such as the
optimization level. The attribute is consumed when initializing the lowering
context to populate the code generation options. CIRGen is updated to generate
and attach this attribute to the module op.