[CIR] Remove the !cir.void return type for functions returning void
C/C++ functions returning void had an explicit !cir.void return type while not
having any returned value, which was breaking a lot of MLIR invariants when the
CIR dialect is used in a greater context, for example with the inliner.
Now, a C/C++ function returning void has not return type and no return values,
which does not break the MLIR invariant about the same number of return types
and returned values.
This change keeps the same parsing/pretty-printed syntax as before for
compatibility.