[CIR][cir-translate] Support specifying target for `cir-translate` (#1186)
This PR adds a new command line option `--target` to our tool
`cir-translate`. The concrete behaviour of it also depends on the triple
and data layout in the CIR module. See the table in code comments for
details.

The default triple is `x86_64-unknown-linux-gnu` currently.

Some tests are updated with triple and DLTI attribute eliminated
(replaced by an option in RUN line). But still some tests remain
unchanged, primarily because they use `cir-opt` instead.