[CIR][CIRGen][TBAA] Replace hardcoded TBAA names with getTBAAName (#1242)
This patch follows
https://github.com/llvm/clangir/pull/1220#discussion_r1882808223 by
augmenting `CIR_Type` with a new field, `tbaaName`. Specifically, it
enables TableGen support for the `-gen-cir-tbaa-name-lowering` option,
allowing for the generation of `getTBAAName` functions based on the
`tbaaName`. This enhancement enables us to replace the hardcoded TBAA
names in the `getTypeName` function with the newly generated
`getTBAAName`.