[CIR] Lower `cir.bool` to i1 (#1158)
This PR changes changes the lowering of `cir.bool` to `i1` in both
DorectToLLVM and ThroughMLIR. This dramatically simplifies the lowering
logic of most operations and the lowered code itself as it naturally
uses `i1` for anything boolean.

The change involves separating between type lowering when scalars are
involved and when memory is involved. This is a pattern that was
inspired by clang's codegen which directly emits `i1` from the AST
without intermediate higher level representation like CIR has.

This also paves the way to more complex lowerings that are implemented
in clang codegen through the three primitives added here: `Convert Type
For Memory`, `Emit For Memory` and `Emit To Memory`. They are used in
clang for non-trivial types like bitints but also extensible vectors.