[CIR] Remove return !cir.void from IR and textual representation
C/C++ functions returning void had an explicit !cir.void return type while not
having any returned value, which was breaking a lot of MLIR invariants when the
CIR dialect is used in a greater context, for example with the inliner.

Now, a C/C++ function returning void has no return type and no return values,
which does not break the MLIR invariant about the same number of return types
and returned values.

This change does not keeps the same parsing/pretty-printed syntax as before for
compatibility like in https://github.com/llvm/clangir/pull/1203 because it
requires some new features from the MLIR parser infrastructure itself, which is
not great.

For now use a list of return types.