[CIR] Add support for array new
This change adds initial support for array new expressions where the
array size is constant. This includes implementing the array cookie
generation in the CIR CXXABI code, but only for the base Itanium ABI.

A subsequent patch will create a subclass of CIRGenItaniumCXXABI to provide
the proper array cookie support for AppleARM64 targets.

This patch does not support variable array sizes or asan handling, both of
which are marked as "NYI".